% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEcopathCatchTotal.R
\name{addEcopathCatchTotal}
\alias{addEcopathCatchTotal}
\title{Sets up gear parameters from Ecopath catch data}
\usage{
addEcopathCatchTotal(params, ecopath_catch)
}
\arguments{
\item{params}{A MizerParams object}

\item{ecopath_catch}{The Ecopath Catch data frame}
}
\description{
Currently this sets up only a single gear for the total catch. The
\code{yield_observed} gear parameter is set to the total catch for each species.
The \code{catchability} is set to the ratio of the observed yield to the observed
biomass. The selectivity of the gear is assumed to be sigmoidal with 50\% of
fish selected at maturity size and 25\% selected at 90\% of maturity size
(measured in weight). Fishing is turned on with an initial effort of 1. The
result will therefore not be at steady state yet.
}
