% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getTotalProduction}
\alias{getTotalProduction}
\title{Get total production}
\usage{
getTotalProduction(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of gonadic production for each species
}
\description{
For each species returns the rate at which biomass is produced. This includes
both somatic and gonadic production.
}
\seealso{
Other rate functions: 
\code{\link{getConsumption}()},
\code{\link{getDietMatrix}()},
\code{\link{getEcotrophicEfficiency}()},
\code{\link{getGonadicProduction}()},
\code{\link{getM0B}()},
\code{\link{getM2B}()},
\code{\link{getMetabolicRespiration}()},
\code{\link{getOffspringProduction}()},
\code{\link{getProduction}()},
\code{\link{getReproductiveLoss}()},
\code{\link{getRespiration}()},
\code{\link{getSomaticProduction}()},
\code{\link{getUnassimilated}()},
\code{\link{getZB}()}
}
\concept{rate functions}
