% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_params.R
\name{update_params}
\alias{update_params}
\title{Function that updates params object with new parameter values}
\usage{
update_params(params, species = 1, pars, biomass, w_select)
}
\arguments{
\item{params}{The MizerParams object to update}

\item{species}{The species to update. By default the first species in the
model.}

\item{pars}{A named numeric vector of parameter values}

\item{biomass}{The observed biomass of the species in the range of the model
specified by \code{w_select}.}

\item{w_select}{A logical vector indicating which weight bins were used in
the likelihood calculation}
}
\value{
The updated MizerParams object
}
\description{
This function currently updates the gear selectivity parameters \code{l50} and
\verb{l25, the catchability, the steepness }U\verb{of the maturity ogive and the coefficient}M` of the power-law mortality rate. It then recalculates the
steady state of the model and rescales it to match the observed biomass.
}
